requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
plotly>=5.17.0
scikit-learn>=1.3.0
google-generativeai>=0.3.0



